Refactor: Consolidate emotion analysis components
Integrates EEG and text emotion analysis into a unified engine. Replaces Goertzel with FFT for EEG bandpower calculation. Adds multiprocessing option for EEG analysis. Includes Korean text emotion processing with fallback mechanism. Updates tests and benchmark template.
