+++ emotion_frequency_analyzer.py
@@ -1,72 +1,23 @@
-emotion_frequency_analyzer.py — 핵심 클래스 EmotionFrequencyAnalyzer (Goertzel, 슬라이딩 윈도우 2048, 칼만 필터, SNR 검사, SIMD-최적화된 numpy 벡터화 등 포함)
-
-tests/test_frequency_analyzer.py — pytest 기반 단위 테스트 (합성 신호로 감정검증)
-
-performance_benchmark.json — 성능 벤치마크 예시(측정 포맷, 실제 환경에서 실행해 채우실 수 있도록 구조 제공)
-
-주파수 밴드: delta(0.5–4), theta(4–8), alpha(8–13), beta(13–30), gamma(30–100)Hz
-
-EEG→감정 매핑(중심값 사용):
-
-joy: alpha=+0.35, beta=+0.25
-
-sadness: alpha=-0.25, theta=+0.20
-
-anger: beta=+0.40, gamma=+0.25
-
-
-계산: Goertzel (특정 주파수군에 대해 벡터화로 빠르게 파워 계산 — numpy 사용으로 SIMD 활용 가능)
-
-윈도우: 슬라이딩 윈도우 길이 2048 샘플, hop configurable (기본 1024)
-
-잡음 처리: 1D Kalman filter 채널별 적용, SNR 계산 후 품질 체크(요구: SNR > 20 dB)
-
-출력: 감정 점수(softmax 정규화), 신호 품질 메타데이터, 밴드별 파워 등
-
-
-
-
-
----
-
-1) emotion_frequency_analyzer.py
-
-# emotion_frequency_analyzer.py
+#!/usr/bin/env python3
 """
-EmotionFrequencyAnalyzer
-- Implements EEG band-power extraction using Goertzel algorithm (vectorized),
-  sliding windows, channel-wise Kalman filtering, SNR quality checks.
-- Maps band features to emotion scores using supplied coefficients.
-
-API:
-    analyzer = EmotionFrequencyAnalyzer(sampling_rate=512, window_size=2048, hop_size=1024)
-    result = analyzer.analyze_eeg_to_emotion(eeg_data, sampling_rate=512)
-Where eeg_data can be:
-    - numpy array shape (n_channels, n_samples)
-    - dict {channel_name: samples_list/np.array}
-Returns:
-{
-  'emotion_vector': {'joy':..., 'sadness':..., 'anger':..., 'neutral':...},
-  'primary_emotion': 'joy',
-  'band_powers': {...},
-  'snr_db': <float>,
-  'signal_quality': 'good'|'poor'|'reject',
-  'confidence': 0.0-1.0,
-  'metadata': { ... }
-}
+emotion_frequency_analyzer.py
+- EEG band-power extraction via Goertzel, sliding windows, Kalman denoising, SNR checks
+- Maps band features to emotion scores (joy, sadness, anger, neutral)
 """
+
 from typing import Dict, Tuple, Union, List
 import numpy as np
-import math
 import time
 
 EPS = 1e-12
 
+
 class Kalman1D:
     """Simple 1D Kalman filter for denoising each channel (scalar process)."""
-    def __init__(self, q=1e-5, r=1e-2, initial_x=0.0, initial_p=1.0):
-        self.q = q  # process variance
-        self.r = r  # measurement variance
+
+    def __init__(self, q: float = 1e-5, r: float = 1e-2, initial_x: float = 0.0, initial_p: float = 1.0) -> None:
+        self.q = q
+        self.r = r
         self.x = initial_x
         self.p = initial_p
 
@@ -74,35 +25,24 @@ class Kalman1D:
         x = self.x
         p = self.p
         out = np.empty_like(measurements)
-        for i, z in enumerate(measurements):
-            # Predict
+        for _, z in enumerate(measurements):
             p = p + self.q
-            # Update
             k = p / (p + self.r)
             x = x + k * (z - x)
             p = (1 - k) * p
-            out[i] = x
-        # Save state
+            out[_] = x
         self.x = x
         self.p = p
         return out
 
 
 def goertzel_power_block(x: np.ndarray, freqs: np.ndarray, fs: float) -> np.ndarray:
-    """
-    Vectorized Goertzel for a 1D signal x and multiple target frequencies.
-    x : 1D array (n_samples,)
-    freqs: 1D array of frequencies to evaluate
-    returns: power for each frequency (len(freqs),)
-    """
+    """Vectorized Goertzel for a 1D signal x and multiple target frequencies."""
     n = x.shape[0]
     k = np.round((n * freqs) / fs).astype(int)
-    # Angular coefficient
     omega = (2.0 * np.pi * k) / n
     cos_omega = np.cos(omega)
     coeff = 2.0 * cos_omega
-
-    # Vectorized Goertzel approach: iterate frequencies but with numpy ops
     powers = np.zeros_like(freqs, dtype=np.float64)
     for i, c in enumerate(coeff):
         s_prev = 0.0
@@ -111,7 +51,6 @@ def goertzel_power_block(x: np.ndarray, freqs: np.ndarray, fs: float) -> np.ndar
             s = sample + c * s_prev - s_prev2
             s_prev2 = s_prev
             s_prev = s
-        # compute real and imaginary parts
         real = s_prev - s_prev2 * cos_omega[i]
         imag = s_prev2 * np.sin(omega[i])
         powers[i] = real * real + imag * imag
@@ -127,27 +66,23 @@ class EmotionFrequencyAnalyzer:
         snr_threshold_db: float = 20.0,
         kalman_q: float = 1e-5,
         kalman_r: float = 1e-2,
-    ):
+    ) -> None:
         self.fs = sampling_rate
         self.window_size = window_size
         self.hop_size = hop_size
         self.snr_threshold_db = snr_threshold_db
-        # Frequency bands per request
-        self.frequency_bands = {
+        self.frequency_bands: Dict[str, Tuple[float, float]] = {
             'delta': (0.5, 4.0),
             'theta': (4.0, 8.0),
             'alpha': (8.0, 13.0),
             'beta': (13.0, 30.0),
-            'gamma': (30.0, 100.0)
+            'gamma': (30.0, 100.0),
         }
-        # Emotion coefficients (central values from user)
-        # Only include the explicitly requested emotions & bands
-        self.emotion_coeffs = {
+        self.emotion_coeffs: Dict[str, Dict[str, float]] = {
             'joy': {'alpha': 0.35, 'beta': 0.25},
             'sadness': {'alpha': -0.25, 'theta': 0.20},
-            'anger': {'beta': 0.40, 'gamma': 0.25}
+            'anger': {'beta': 0.40, 'gamma': 0.25},
         }
-        # Kalman params
         self.kalman_q = kalman_q
         self.kalman_r = kalman_r
 
@@ -156,10 +91,9 @@ class EmotionFrequencyAnalyzer:
             channels = list(eeg_input.keys())
             arrs = [np.asarray(eeg_input[ch], dtype=np.float64) for ch in channels]
             n_samples = max(a.shape[0] for a in arrs)
-            # pad shorter channels with zeros
             data = np.zeros((len(arrs), n_samples), dtype=np.float64)
             for i, a in enumerate(arrs):
-                data[i, :a.shape[0]] = a
+                data[i, : a.shape[0]] = a
             return data, channels
         elif isinstance(eeg_input, np.ndarray):
             if eeg_input.ndim == 1:
@@ -168,158 +102,103 @@ class EmotionFrequencyAnalyzer:
                 channels = [f'ch{i}' for i in range(eeg_input.shape[0])]
                 return eeg_input.astype(np.float64), channels
             else:
-                raise ValueError("eeg_input numpy array must be 1D or 2D (channels x samples).")
+                raise ValueError('eeg_input numpy array must be 1D or 2D (channels x samples).')
         else:
-            raise TypeError("eeg_input must be numpy.ndarray or dict")
+            raise TypeError('eeg_input must be numpy.ndarray or dict')
 
     def _apply_kalman(self, data: np.ndarray) -> np.ndarray:
-        """Apply channel-wise 1D Kalman filter"""
         filtered = np.empty_like(data)
         for i in range(data.shape[0]):
-            kf = Kalman1D(q=self.kalman_q, r=self.kalman_r, initial_x=float(np.median(data[i,:])))
+            kf = Kalman1D(q=self.kalman_q, r=self.kalman_r, initial_x=float(np.median(data[i, :])))
             filtered[i, :] = kf.filter(data[i, :])
         return filtered
 
     def _calc_snr_db(self, signal: np.ndarray, noise_estimate: np.ndarray) -> float:
-        # Compute power
         ps = np.mean(signal ** 2) + EPS
         pn = np.mean(noise_estimate ** 2) + EPS
         snr = 10.0 * np.log10(ps / pn)
         return float(snr)
 
-    def _band_power_goertzel(self, segment: np.ndarray, band: Tuple[float, float], fs: float, n_freqs:int=5) -> float:
-        """
-        Compute approximate band power by evaluating n_freqs evenly spaced frequencies inside band
-        using Goertzel and averaging.
-        """
+    def _band_power_goertzel(self, segment: np.ndarray, band: Tuple[float, float], fs: float, n_freqs: int = 5) -> float:
         low, high = band
         if high <= low:
             return 0.0
-        freqs = np.linspace(max(0.1, low), min(fs/2 - 0.1, high), n_freqs)
+        freqs = np.linspace(max(0.1, low), min(fs / 2 - 0.1, high), n_freqs)
         powers = goertzel_power_block(segment, freqs, fs)
-        # band power estimate: mean of frequency powers
         return float(np.mean(powers) + EPS)
 
     def _extract_band_features(self, data_segment: np.ndarray) -> Dict[str, float]:
-        """
-        data_segment: 1D per-channel averaged across channels OR single channel segment.
-        We will compute band powers per band and return dict.
-        """
-        band_powers = {}
-        # vectorized in the sense we compute per band using goertzel on the averaged signal
-        # average across channels (SIMD benefit from numpy)
         if data_segment.ndim == 2:
-            # channels x samples -> average to 1D
             avg = np.mean(data_segment, axis=0)
         else:
             avg = data_segment
+        band_powers: Dict[str, float] = {}
         for band_name, (low, high) in self.frequency_bands.items():
             band_powers[band_name] = self._band_power_goertzel(avg, (low, high), self.fs, n_freqs=5)
         return band_powers
 
-    def _map_bands_to_emotions(self, band_features: Dict[str, float]) -> Tuple[Dict[str, float], float]:
-        """
-        Map band_features (powers) to emotion raw scores using linear combination
-        using self.emotion_coeffs. Returns normalized emotion vector and raw primary score.
-        """
-        # Normalize band features to 0..1 by dividing by sum
+    def _map_bands_to_emotions(self, band_features: Dict[str, float]) -> Tuple[Dict[str, float], str]:
         total = sum(band_features.values()) + EPS
         norm = {k: v / total for k, v in band_features.items()}
-
-        raw_scores = {}
+        raw_scores: Dict[str, float] = {}
         for emotion, coeffs in self.emotion_coeffs.items():
             s = 0.0
             for band, coeff in coeffs.items():
-                # Some coefficients reference 'gamma' or 'theta' etc.
                 band_val = norm.get(band, 0.0)
                 s += coeff * band_val
             raw_scores[emotion] = s
-
-        # Add a 'neutral' baseline (small positive to avoid all zeros)
         raw_scores['neutral'] = 0.0
-
-        # ReLU to avoid negative raw-values harming softmax
         relu_scores = {k: max(0.0, v) for k, v in raw_scores.items()}
-
-        # Softmax normalize
         vals = np.array(list(relu_scores.values()), dtype=np.float64)
         if vals.sum() <= EPS:
-            # fallback: small uniform vector
             soft = np.ones_like(vals) / len(vals)
         else:
             exps = np.exp(vals - np.max(vals))
             soft = exps / (exps.sum() + EPS)
-
         emotion_vector = {k: float(soft[i]) for i, k in enumerate(relu_scores.keys())}
-        # primary emotion is argmax
         primary = max(emotion_vector.items(), key=lambda kv: kv[1])[0]
         return emotion_vector, primary
 
     def analyze_eeg_to_emotion(self, eeg_input: Union[np.ndarray, Dict[str, np.ndarray]], sampling_rate: int = None) -> Dict:
-        """
-        Top-level method to analyze EEG buffer into emotion vector.
-
-        EEG input is processed in sliding windows. The method computes:
-          - Kalman-filtered signal
-          - SNR estimate (global)
-          - Band powers (averaged over channels)
-          - Emotion mapping
-
-        Returns final aggregated emotion vector and metadata.
-        """
         t0 = time.time()
         if sampling_rate is not None and sampling_rate != self.fs:
-            raise ValueError(f"sampling_rate mismatch: analyzer fs={self.fs}, provided={sampling_rate}")
+            raise ValueError(f'sampling_rate mismatch: analyzer fs={self.fs}, provided={sampling_rate}')
 
         data, channel_names = self._ensure_numpy(eeg_input)
         n_ch, n_samples = data.shape
 
-        # Signal quality check: quick SNR estimate using simple median-based noise estimate
-        raw_power = np.mean(data ** 2)
-        # estimate noise as high-frequency residual using a simple diff operator
+        _ = np.mean(data ** 2)
         noise_est = np.mean(np.diff(data, axis=1) ** 2) if n_samples > 1 else 1e-6
-        # For SNR, use average across channels
         snr_db = self._calc_snr_db(data, np.sqrt(noise_est) * np.ones_like(data))
-
-        # If below threshold, flag; try filtering anyway
         signal_quality = 'good' if snr_db > self.snr_threshold_db else ('poor' if snr_db > (self.snr_threshold_db - 6) else 'reject')
 
-        # Apply Kalman filter channel-wise (can be expensive; windows are used later)
         try:
             filtered = self._apply_kalman(data)
         except Exception:
             filtered = data.copy()
 
-        # Sliding window aggregation: compute band features per window and average
         band_accum = {b: 0.0 for b in self.frequency_bands.keys()}
         window_count = 0
         start = 0
         while start + self.window_size <= n_samples:
-            seg = filtered[:, start:start + self.window_size]
+            seg = filtered[:, start : start + self.window_size]
             band_feats = self._extract_band_features(seg)
             for k, v in band_feats.items():
                 band_accum[k] += v
             window_count += 1
             start += self.hop_size
         if window_count == 0:
-            # not enough samples for a single window; process entire signal as one window
             band_feats = self._extract_band_features(filtered)
             for k, v in band_feats.items():
                 band_accum[k] += v
             window_count = 1
 
-        # average band powers
         band_powers = {k: (v / window_count) for k, v in band_accum.items()}
-
-        # Map to emotions
         emotion_vector, primary = self._map_bands_to_emotions(band_powers)
 
-        # Confidence heuristic: combination of SNR and primary magnitude
         primary_score = emotion_vector.get(primary, 0.0)
-        snr_factor = min(1.0, max(0.0, (snr_db - self.snr_threshold_db + 6.0) / 12.0))  # scaled
+        snr_factor = min(1.0, max(0.0, (snr_db - self.snr_threshold_db + 6.0) / 12.0))
         confidence = float(primary_score * 0.7 + snr_factor * 0.3)
-
-        # If signal rejected, set low confidence
         if signal_quality == 'reject':
             confidence = min(confidence, 0.25)
 
@@ -329,7 +208,7 @@ class EmotionFrequencyAnalyzer:
             'n_samples': n_samples,
             'window_size': self.window_size,
             'hop_size': self.hop_size,
-            'processing_time_ms': (time.time() - t0) * 1000.0
+            'processing_time_ms': (time.time() - t0) * 1000.0,
         }
 
         return {
@@ -339,163 +218,16 @@ class EmotionFrequencyAnalyzer:
             'snr_db': float(snr_db),
             'signal_quality': signal_quality,
             'confidence': float(confidence),
-            'metadata': metadata
+            'metadata': metadata,
         }
 
 
 if __name__ == '__main__':
-    # quick demo run (not a test)
-    import numpy as np
     fs = 512
-    t = np.arange(0, 5.0, 1.0/fs)
-    # synthetic alpha(10Hz) + beta(20Hz) -> joy
-    sig = 0.8*np.sin(2*np.pi*10*t) + 0.5*np.sin(2*np.pi*20*t) + 0.05*np.random.randn(t.size)
-    eeg = np.stack([sig, sig*0.9])
+    t = np.arange(0, 5.0, 1.0 / fs)
+    sig = 0.8 * np.sin(2 * np.pi * 10 * t) + 0.5 * np.sin(2 * np.pi * 20 * t) + 0.05 * np.random.randn(t.size)
+    eeg = np.stack([sig, sig * 0.9])
     analyzer = EmotionFrequencyAnalyzer(sampling_rate=fs)
     res = analyzer.analyze_eeg_to_emotion(eeg)
     print(res)
 
-
----
-
-2) tests/test_frequency_analyzer.py
-
-# tests/test_frequency_analyzer.py
-import numpy as np
-import pytest
-from emotion_frequency_analyzer import EmotionFrequencyAnalyzer
-
-def synth_signal(fs, duration, components):
-    """
-    components: list of tuples (freq, amplitude)
-    returns 1D array
-    """
-    t = np.arange(0, duration, 1.0/fs)
-    sig = np.zeros_like(t)
-    for f, a in components:
-        sig += a * np.sin(2 * np.pi * f * t)
-    # small gaussian noise to make SNR realistic
-    sig += 0.01 * np.random.randn(t.size)
-    return sig
-
-@pytest.fixture(scope="module")
-def analyzer():
-    return EmotionFrequencyAnalyzer(sampling_rate=512, window_size=2048, hop_size=1024)
-
-def test_joy_detection(analyzer):
-    fs = 512
-    # boost alpha (10Hz) and beta (20Hz) -> should map to joy
-    sig = synth_signal(fs, duration=6.0, components=[(10.0, 1.0), (20.0, 0.6)])
-    eeg = np.stack([sig, 0.9*sig])  # two channels
-    res = analyzer.analyze_eeg_to_emotion(eeg)
-    assert res['snr_db'] > 10.0  # ensure signal is reasonably clean
-    assert 'joy' in res['emotion_vector']
-    # primary emotion should be joy or at least highest score among defined emotions
-    primary = res['primary_emotion']
-    assert primary == 'joy'
-    assert res['confidence'] > 0.3
-
-def test_sadness_detection(analyzer):
-    fs = 512
-    # boost theta (6Hz) and negative alpha component (we simulate lower alpha by smaller amp)
-    sig = synth_signal(fs, duration=6.0, components=[(6.0, 1.0), (10.0, 0.2)])
-    eeg = np.stack([sig, sig*1.05])
-    res = analyzer.analyze_eeg_to_emotion(eeg)
-    # primary should be sadness (theta dominant)
-    assert res['primary_emotion'] == 'sadness'
-
-def test_anger_detection(analyzer):
-    fs = 512
-    # boost beta (20Hz) and gamma (40Hz)
-    sig = synth_signal(fs, duration=6.0, components=[(20.0, 1.0), (40.0, 0.8)])
-    eeg = np.stack([sig, 0.8*sig])
-    res = analyzer.analyze_eeg_to_emotion(eeg)
-    assert res['primary_emotion'] == 'anger'
-
-def test_low_snr_rejection(analyzer):
-    fs = 512
-    # produce noisy signal low amplitude -> should reduce confidence and possibly 'reject'
-    t = np.arange(0, 3.0, 1.0/fs)
-    noise = 0.5 * np.random.randn(t.size)
-    eeg = np.stack([noise, 1.0*noise])
-    res = analyzer.analyze_eeg_to_emotion(eeg)
-    # snr likely to be low -> signal_quality 'reject' or 'poor'
-    assert res['signal_quality'] in ('reject', 'poor')
-    assert res['confidence'] <= 0.5
-
-설치/실행 방법 (간단):
-
-# 가상환경 권장
-pip install numpy pytest
-# 루트에서
-pytest -q
-
-
----
-
-3) performance_benchmark.json (템플릿 및 예시 결과)
-
-아래 JSON은 벤치마크 결과를 기록하는 표준 포맷입니다. 실제 환경(하드웨어, cpu, 샘플 수)에 따라 값을 채우시고, CI에서 자동 측정하도록 하시면 좋습니다.
-
-{
-  "benchmark_run": {
-    "timestamp": "2025-09-08T00:00:00Z",
-    "hardware": {
-      "cpu": "Intel(R) Xeon(R) or dev machine (specify)",
-      "num_cores": 8,
-      "ram_gb": 32
-    },
-    "settings": {
-      "sampling_rate": 512,
-      "window_size": 2048,
-      "hop_size": 1024,
-      "n_channels": 8,
-      "duration_seconds": 60
-    }
-  },
-  "metrics": {
-    "per_window": {
-      "avg_processing_time_ms": 12.3,
-      "p50_processing_time_ms": 11.0,
-      "p95_processing_time_ms": 22.6,
-      "p99_processing_time_ms": 48.7
-    },
-    "throughput": {
-      "windows_per_second": 80.5,
-      "concurrent_sessions_supported": 120
-    },
-    "memory_usage_mb": {
-      "baseline": 85.7,
-      "peak": 142.3
-    },
-    "cpu_usage_percent": {
-      "avg": 35.4,
-      "peak": 78.1
-    },
-    "notes": "Example numbers. Run on target hardware and replace these values. Use -O numpy builds and MKL/BLAS for SIMD acceleration."
-  }
-}
-
-> 벤치마크 가이드:
-
-실제 측정은 time.perf_counter()로 각 윈도우 처리 시간(Goertzel+Kalman+map 단계)을 측정한 뒤 요약 통계 제공하세요.
-
-numpy가 Intel MKL/BLIS 등으로 빌드되어 있으면 SIMD가 활성화되어 실시간 처리 성능이 향상됩니다.
-
-대량 동시 처리(여러 세션)를 위해서는 윈도우/채널 정규화 및 batch 처리(여러 채널을 한 번에 처리)로 CPU 캐시/벡터 유틸리티를 최대화 하세요.
-
-
-
-
-
----
-
-구현/테스트 관련 추가 메모 (짧게)
-
-Goertzel: 구현은 각 타겟 주파수별 반복 루프가 있으나, 우리는 윈도우 내 여러 주파수(밴드당 5개)만 평가하므로 전체 FFT보다 훨씬 효율적입니다. numpy 벡터화를 활용하면 내부 루프(샘플 반복)는 파이썬 루프로 남지만 window_size=2048 정도에서 비용-이익이 적절합니다. 대량 배치(여러 창, 여러 채널)를 한 번에 처리하려면 Goertzel를 C/Numba로 JIT 하거나 FFT 기반 밴드파워를 병렬화하는 것도 고려하세요.
-
-Kalman 필터: 단순 1D 채널별 필터 사용(실시간 일차원 노이즈 억제). 보다 정교한 아티팩트 제거(눈 깜빡임, 근전도 등)는 ICA/ASR 같은 추가 파이프라인을 권장합니다.
-
-SNR 기준: 요구대로 SNR > 20 dB를 목표로 하였고, 품질 레이블('good'/'poor'/'reject')와 confidence 스케일에 반영했습니다.
-
-확장: multi-emotion(7개)로 확장 시 self.emotion_coeffs에 추가 계수를 넣어 동일 파이프라인 사용 가능.
