+++ korean_emotion_processor.py
@@ -1,11 +1,9 @@
-```python
 #!/usr/bin/env python3
 """
 korean_emotion_processor.py
-- KoreanEmotionProcessor 클래스: 한국어 형태소 분석, 어미별 감정 점수 계산, 반어법 감지.
+- KoreanEmotionProcessor: 한국어 형태소 분석, 어미별 감정 점수 계산, 반어법 감지.
 - text_to_emotion_music: 텍스트 → 형태소 추출 → 어미 매핑 → 감정 강도 → 음악 코드 생성.
-- Konlpy 대체: 환경 제한으로 단순 문자열 매칭 사용 (konlpy import 시도 후 fallback).
-- 성능: 1000자 < 30ms, 메모리 < 100MB 목표 (benchmark에서 검증).
+- Konlpy 시도 후 실패 시 간단 정규식 기반 fallback 사용.
 """
 
 import time
@@ -17,21 +15,22 @@ from typing import Dict, List
 try:
     from konlpy.tag import Komoran
     KONLPY_AVAILABLE = True
-except ImportError:
+except Exception:
     KONLPY_AVAILABLE = False
-    print("⚠️ Konlpy not available, using simple string matching for morpheme analysis.")
+
 
 class KoreanEmotionProcessor:
     """
     한국어 감정 처리기: 형태소 분석, 어미 감정 매핑, 반어법 감지.
     """
+
     def __init__(self):
         # 제공된 한국어 데이터
         self.ending_emotions = {
             '네요': {'emotion': '놀람/깨달음', 'intensity': '중'},
             '군요': {'emotion': '발견/인정', 'intensity': '중'},
             '거든요': {'emotion': '정당화/단호', 'intensity': '중'},
-            '잖아요': {'emotion': '공유지식', 'intensity': '중'}
+            '잖아요': {'emotion': '공유지식', 'intensity': '중'},
         }
         # 반어법 키워드
         self.sarcasm_keywords = ["참", "그래 잘했어"]
@@ -43,34 +42,33 @@ class KoreanEmotionProcessor:
             '발견/인정': 'G Major',
             '정당화/단호': 'D Minor',
             '공유지식': 'F Major',
-            '반어법': 'E7'  # 긴장/반전 코드
+            '반어법': 'E7',  # 긴장/반전 코드
         }
         if KONLPY_AVAILABLE:
             self.komoran = Komoran()
 
     def extract_morphemes(self, text: str) -> List[str]:
-        """형태소 추출: Konlpy 사용 또는 fallback 문자열 split."""
+        """형태소 추출: Konlpy 사용 또는 fallback 정규식.
+
+        Fallback에서는 문장 내 어디서든 어미를 탐지합니다(문장부호와 무관).
+        """
         if KONLPY_AVAILABLE:
-            return [morph for morph, tag in self.komoran.pos(text) if tag in ['EF', 'EC']]  # 어미 태그
-        else:
-            # Fallback: 정규식으로 어미 추출 (간단 패턴)
-            endings = re.findall(r'(네요|군요|거든요|잖아요)$', text)
-            return endings
+            return [morph for morph, tag in self.komoran.pos(text) if tag in ('EF', 'EC')]
+        # Fallback: 간단 어미 매칭 (문장 끝 고정 아님)
+        return re.findall(r'(네요|군요|거든요|잖아요)', text)
 
     def detect_sarcasm(self, text: str) -> bool:
         """반어법 감지: 키워드 매칭."""
-        for kw in self.sarcasm_keywords:
-            if kw in text:
-                return True
-        return False
+        return any(keyword in text for keyword in self.sarcasm_keywords)
 
     def calculate_emotion_scores(self, morphemes: List[str]) -> Dict[str, float]:
         """어미별 감정 점수 계산."""
-        scores = {}
+        scores: Dict[str, float] = {}
         for morph in morphemes:
             if morph in self.ending_emotions:
-                emotion = self.ending_emotions[morph]['emotion']
-                intensity = self.intensity_map.get(self.ending_emotions[morph]['intensity'], 0.5)
+                info = self.ending_emotions[morph]
+                emotion = info['emotion']
+                intensity = self.intensity_map.get(info['intensity'], 0.5)
                 scores[emotion] = intensity
         return scores
 
@@ -80,22 +78,30 @@ class KoreanEmotionProcessor:
         morphemes = self.extract_morphemes(text)
         is_sarcasm = self.detect_sarcasm(text)
         scores = self.calculate_emotion_scores(morphemes)
-        
+
         # 감정 강도 계산 (평균)
-        total_intensity = sum(scores.values()) / max(1, len(scores))
-        primary_emotion = max(scores, key=scores.get) if scores else '중립'
-        
+        if scores:
+            total_intensity = sum(scores.values()) / max(1, len(scores))
+            primary_emotion = max(scores, key=scores.get)
+        else:
+            total_intensity = 0.0
+            primary_emotion = '중립'
+
         # 반어법 적용 (감정 반전)
         if is_sarcasm:
             primary_emotion = '반어법'
-            total_intensity *= -1  # 부정으로 flip
-        
+            total_intensity *= -1
+
         # 음악 코드 생성
-        chord = self.emotion_to_chord.get(primary_emotion, 'C Minor')  # 기본 부정 코드
-        
-        processing_time = (time.time() - start_time) * 1000  # ms
-        memory_usage = sys.getsizeof(text) + sys.getsizeof(morphemes) + sys.getsizeof(scores)  # approximate bytes
-        
+        chord = self.emotion_to_chord.get(primary_emotion, 'C Minor')
+
+        processing_time_ms = (time.time() - start_time) * 1000.0
+        memory_usage_bytes = (
+            sys.getsizeof(text)
+            + sys.getsizeof(morphemes)
+            + sys.getsizeof(scores)
+        )
+
         return {
             'morphemes': morphemes,
             'emotions': scores,
@@ -103,67 +109,12 @@ class KoreanEmotionProcessor:
             'intensity': total_intensity,
             'music_chord': chord,
             'is_sarcasm': is_sarcasm,
-            'processing_time_ms': processing_time,
-            'memory_usage_bytes': memory_usage
+            'processing_time_ms': processing_time_ms,
+            'memory_usage_bytes': memory_usage_bytes,
         }
 
-# ... (통합 클래스 등, 이전 코드와 병합 가능)
-```
 
-```python
-#!/usr/bin/env python3
-"""
-tests/test_korean.py
-- KoreanEmotionProcessor 단위 테스트.
-- 예시 텍스트로 어미 추출, 감정 점수, 반어법 감지, 음악 변환 검증.
-"""
-
-import unittest
-from korean_emotion_processor import KoreanEmotionProcessor  # 가정: 동일 디렉토리
-
-class TestKoreanEmotionProcessor(unittest.TestCase):
-    def setUp(self):
-        self.processor = KoreanEmotionProcessor()
-
-    def test_extract_morphemes(self):
-        text = "이게 재미있네요."
-        morphemes = self.processor.extract_morphemes(text)
-        if KONLPY_AVAILABLE:
-            self.assertIn('네요', morphemes)
-        else:
-            self.assertIn('네요', morphemes)
-
-    def test_detect_sarcasm(self):
-        text1 = "참 잘했어."
-        text2 = "좋은 아침입니다."
-        self.assertTrue(self.processor.detect_sarcasm(text1))
-        self.assertFalse(self.processor.detect_sarcasm(text2))
-
-    def test_calculate_emotion_scores(self):
-        morphemes = ['네요', '군요']
-        scores = self.processor.calculate_emotion_scores(morphemes)
-        self.assertIn('놀람/깨달음', scores)
-        self.assertEqual(scores['놀람/깨달음'], 0.5)
-
-    def test_text_to_emotion_music(self):
-        text = "이게 재미있네요. 참 잘했어."
-        result = self.processor.text_to_emotion_music(text)
-        self.assertIn('emotions', result)
-        self.assertTrue(result['is_sarcasm'])
-        self.assertLess(result['processing_time_ms'], 30)  # 1000자 미만 테스트
-
-if __name__ == '__main__':
-    unittest.main()
-```
-
-```json
-{
-  "benchmark_results": {
-    "test_text_length": 1000,
-    "processing_time_ms": 12.34,
-    "memory_usage_mb": 0.05,
-    "meets_requirements": true
-  },
-  "notes": "1000자 텍스트 처리: < 30ms, < 100MB. Fallback 모드 사용 시 더 빠름."
-}
-```
+__all__ = [
+    'KoreanEmotionProcessor',
+    'KONLPY_AVAILABLE',
+]
